pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "client_server_einkaufsliste"
version = "0.1.0"
description = "Einkaufsliste Beispiel mit FastAPI und responsive Client"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "fastapi>=0.95",
  "uvicorn[standard]>=0.22",
  "sqlmodel>=0.0.8",
  "python-jose[cryptography]>=3.3.0",
  # "passlib[bcrypt]==1.7.4",
  "bcrypt>=5.0.0",
  "python-multipart>=0.0.6"
]

[project.optional-dependencies]
dev = [
  "pytest>=7.0",
  "black>=23.0",
  "flake8>=6.0",
  "httpx>=0.24",
  "pre-commit>=3.0"
]

[tool.black]
line-length = 88
target-version = ["py310"]
include = "\\.pyi?$"
exclude = '''
/(
  \.(git|svn|hg)|\..*|__pycache__|build|dist
)/
'''

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-q"
testpaths = ["server/tests"]

[tool.setuptools.packages.find]
where = ["server"]
include = ["src*"]

